producer/requirements.txt
kafka-python==2.0.2
Flask==1.1.2
